backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6
opencv-python-headless>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0,<2.0.0
tensorflow-cpu>=2.13.0
fer>=22.5.1
